Year 2015 Day 01 (#11)
* Changed al ulong to long

* Day 01 Problem 01

* Day 01 Problem 2

+++ AdventOfCode/AdventOfCode2020/Day06.cs
@@ -7,16 +7,16 @@ namespace AdventOfCode2020
 {
 	public class Day06 : Day<string[]>
 	{
-		public override ulong Perform1(string inputString)
+		public override long Perform1(string inputString)
 		{
 			var inputStrings = this.ParseInput(inputString);
-			return (ulong) inputStrings.Select(ParseInput1).Sum();
+			return inputStrings.Select(ParseInput1).Sum();
 		}
 
-		public override ulong Perform2(string inputString)
+		public override long Perform2(string inputString)
 		{
 			var inputStrings = this.ParseInput(inputString);
-			return (ulong) inputStrings.Select(ParseInput2).Sum();
+			return inputStrings.Select(ParseInput2).Sum();
 		}
 
 		protected override string[] ParseInput(string inputString)
